% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{descriptives}
\alias{descriptives}
\title{Descriptive statistics}
\usage{
descriptives(x, name = "", histogram = TRUE, barplot = FALSE,
  densityplot = TRUE, savePDF = FALSE, noPlot = FALSE)
}
\arguments{
\item{x}{.}

\item{name}{.}

\item{histogram}{.}

\item{barplot}{.}

\item{densityplot}{.}

\item{savePDF}{.}

\item{noPlot}{.}
}
\value{
.
}
\description{
An old function I used to explore data. Has to be rewritten and documented.
}
