% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_CI}
\alias{get_CI}
\title{Compute confidence interval}
\usage{
get_CI(x, conf.level = 0.95, dist = "t", incl_n = TRUE,
  incl_mean = TRUE, incl_SD = TRUE, incl_SE = TRUE,
  incl_error = TRUE)
}
\arguments{
\item{x}{A numeric (or logical) vector (such as a variable)}

\item{conf.level}{Confidence level of the interval. Defaults to 0.95.}

\item{dist}{Distribution the confidence interval should be based on. "t" or
"student" (default) for Student's t-distribution, "z" or "normal" for
standard normal distribution.}

\item{incl_n}{A locigal indicating whether the number of observations should
be included in the returned results. Defaults to TRUE.}

\item{incl_mean}{A locigal indicating whether the mean should
be included in the returned results. Defaults to TRUE.}

\item{incl_SD}{A locigal indicating whether the standard deviation should
be included in the returned results. Defaults to TRUE.}

\item{incl_SE}{A locigal indicating whether the standard error should
be included in the returned results. Defaults to TRUE.}

\item{incl_error}{A locigal indicating whether the error based on the
confidence level (i.e., the number that is substracted or added to the mean
lower and upper bounds of the confidence interval) should
be included in the returned results. Defaults to TRUE.}
}
\value{
A numeric vector with at least the lower and upper bounds of the
confidence interval.
}
\description{
Computes the confidence interval based on Student's t-distribution or the
standard normal distribution. Also reports descriptive statistics (such as
the mean). Missings are ignored.
}
\examples{
## The height of ten people in cm
heights <- c(160, 163, 168, 169, 174, 176, 177, 178, 182, 190)

get_CI(heights) # Defaults
get_CI(heights, dist = "normal") # Using standard normal distribution
get_CI(heights, conf.level = 0.9) # 90\%-CI
get_CI(heights, incl_SD = FALSE) # Do not include standard deviation
}
