% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_helpers.R
\name{get_overview}
\alias{get_overview}
\title{Overview over all variables and their labels in a data set}
\usage{
get_overview(x, trunc_varlab = 0, trunc_vallab = 0)
}
\arguments{
\item{x}{A tibble (read in by the haven package).}

\item{trunc_varlab}{The maximum number of characters for variable labels.
Defaults to 0 for no truncation.}

\item{trunc_vallab}{The maximum number of characters for value labels.
Defaults to 0 for no truncation.}
}
\value{
A data frame with variable names, variable labels, and
value labels.
}
\description{
get_overview returns a data.frame with variable names, variable labels, and
value labels. It should be used with data imported by the \link{haven}
package.
}
