% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_helpers.R
\name{get_value_labels}
\alias{get_value_labels}
\title{Get a list of all value labels in a data set}
\usage{
get_value_labels(x, truncate = 0)
}
\arguments{
\item{x}{A tibble read in by the haven package.}

\item{truncate}{The maximum number of characters for each label. Defaults to
0 for no truncation.}
}
\value{
A named list of all value labels.
}
\description{
get_value_labels returns the value labels generated by data import
functions from the haven package.
}
