% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haven_helpers.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Get a list of all variable labels in a data set}
\usage{
get_variable_labels(x, truncate = 0)
}
\arguments{
\item{x}{A tibble read in by the haven package.}

\item{truncate}{The maximum number of characters for each label. Defaults to
0 for no truncation.}
}
\value{
A named list of all variable labels.
}
\description{
get_variable_labels returns the variable labels generated by data import
functions from the \link{haven} package.
}
