% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variables.R
\name{home_dir}
\alias{home_dir}
\title{Get a user's home directory}
\usage{
home_dir()
}
\value{
The path to the users home directory as a string.
}
\description{
home_dir returns a path to the user's home directory. It replaces "~" and
should work in both RStudio and RScript as well as on different OS.
}
