% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variables.R
\name{invert_variable}
\alias{invert_variable}
\title{Invert a variable}
\usage{
invert_variable(x, max_value)
}
\arguments{
\item{x}{A numeric variable.}

\item{max_value}{Optional, the heighest possible value on the variable's
scale.}
}
\value{
The same variable with an inverted scale.
}
\description{
invert_variable returns the variable with an inverted scale. The heighest
possible value for a variable can be provided. If max_value is not provided,
the maximal value is computed from x.
}
\details{
For a variable with values between 1 and 5, 1 will become 5, 2 will become 4,
1 will become 5, and so on.
}
