% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ordinal_alpha}
\alias{ordinal_alpha}
\title{Ordinal alpha}
\usage{
ordinal_alpha(x)
}
\arguments{
\item{x}{A set of variables (either a data frame or a matrix).}
}
\value{
Cronbach's Alpha for ordinal data.
}
\description{
Calculates the ordinal alpha by using polychoric correlation (see Zumbo et
al. 2007).
}
\references{
Zumbo, B. D., Gadermann, A. M., & Zeisser, C. (2007). Ordinal Versions of Coefficients Alpha and Theta for Likert Rating Skales. Journal of Modern Applied Statistical Methods, 6(1), 21–29.
}
