% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variables.R
\name{rescale_variable}
\alias{rescale_variable}
\title{Rescale a variable}
\usage{
rescale_variable(x, min_old, min_new = 0, max_old, max_new = 1)
}
\arguments{
\item{x}{A numeric variable.}

\item{min_old}{The old minimum value for the variable. Default is
the variable's minimum value.}

\item{min_new}{The new minimum value, defaults to 0.}

\item{max_old}{The old maximum value for the variable. Default is the
variable's maximum value.}

\item{max_new}{The new maximum value, defaults to 1.}
}
\value{
The rescaled variable.
}
\description{
rescale_variable rescales a variable from an old range of values (min_old to
max_old) to a new range (min_new to max_new). Default behavior is to rescale
a variable to a new range from 0 to 1 where the minimum observed value is
recoded to 0 and the maximum observed value is recoded to 1.
}
